updated pio repo, fixed cleanup script for couch db v2

+++ couch_cleanup.sh
@@ -28,34 +28,44 @@ curl -XPOST ${PREFIX}_replicate -H 'Content-Type: application/json' -d'{"source"
 echo "Migration of builds completed."
 
 # swap databases
-rm /opt/couchdb/data/managed_builds.backup
+if [[ -f /opt/couchdb/data/managed_builds.backup ]]; then
+  rm /opt/couchdb/data/managed_builds.backup
+fi
 mv /opt/couchdb/data/managed_builds.couch /opt/couchdb/data/managed_builds.backup
 mv /opt/couchdb/data/managed_builds_${DATE}.couch /opt/couchdb/data/managed_builds.couch
 
 curl -XPOST ${PREFIX}_replicate -H 'Content-Type: application/json' -d'{"source":"managed_devices","target":"managed_devices_'${DATE}'", "create_target":true }'
 echo "Migration of devices completed."
 
-rm /opt/couchdb/data/managed_devices.backup
+if [[ -f /opt/couchdb/data/managed_devices.backup ]]; then
+  rm /opt/couchdb/data/managed_devices.backup
+fi
 mv /opt/couchdb/data/managed_devices.couch /opt/couchdb/data/managed_devices.backup
 mv /opt/couchdb/data/managed_devices_${DATE}.couch /opt/couchdb/data/managed_devices.couch
 
 curl -XPOST ${PREFIX}_replicate -H 'Content-Type: application/json' -d'{"source":"managed_logs","target":"managed_logs_'${DATE}'", "create_target":true }'
 echo "Migration of logs completed."
 
-rm /opt/couchdb/data/managed_logs.backup
+if [[ -f /opt/couchdb/data/managed_logs.backup ]]; then
+  rm /opt/couchdb/data/managed_logs.backup
+fi
 mv /opt/couchdb/data/managed_logs.couch /opt/couchdb/data/managed_logs.backup
 mv /opt/couchdb/data/managed_logs_${DATE}.couch /opt/couchdb/data/managed_logs.couch
 
 curl -XPOST ${PREFIX}_replicate -H 'Content-Type: application/json' -d'{"source":"managed_sources","target":"managed_sources_'${DATE}'", "create_target":true }'
 echo "Migration of sources completed."
 
-rm /opt/couchdb/data/managed_sources.backup
+if [[ -f /opt/couchdb/data/managed_sources.backup ]]; then
+  rm /opt/couchdb/data/managed_sources.backup
+fi
 mv /opt/couchdb/data/managed_sources.couch /opt/couchdb/data/managed_sources.backup
 mv /opt/couchdb/data/managed_sources_${DATE}.couch /opt/couchdb/data/managed_sources.couch
 
 curl -XPOST ${PREFIX}_replicate -H 'Content-Type: application/json' -d'{"source":"managed_users","target":"managed_users_'${DATE}'", "create_target":true, "filter":"users/del"}'
 echo "Migration of users (without deletions) completed."
 
-rm /opt/couchdb/data/managed_users.backup
+if [[ -f /opt/couchdb/data/managed_users.backup ]]; then
+  rm /opt/couchdb/data/managed_users.backup
+fi
 mv /opt/couchdb/data/managed_users.couch /opt/couchdb/data/managed_users.backup
 mv /opt/couchdb/data/managed_users_${DATE}.couch /opt/couchdb/data/managed_users.couch
